requirements.txt
flask
flask_migrate
flask_sqlalchemy
psycopg2-binary
requests
pytest
pytest-flask
sqlalchemy